extracts to speckle settings into manager singleton
